Add -pthread flag
Hello, I could not compile Koi on openSUSE Tumbleweed due to `undefined reference to symbol 'pthread_create@@GLIBC_2.2.5'`. This means that `pthread` was [missing as a link dependency](https://stackoverflow.com/questions/34143265/undefined-reference-to-symbol-pthread-createglibc-2-2-5). To fix it, I added to the `CMAKE_CXX_FLAGS` the `-pthread` flag (pthread, unlike other libraries, is not simply added as an external library, since there are [differences between using -pthread and -lpthread](https://stackoverflow.com/questions/23250863/difference-between-pthread-and-lpthread-while-compiling)). This small fix makes the build finish correctly.