Replaces instance check with shared memory; Removes QT Network dependency (#113)
Koi checks, if another instance is running.

Current behaviour will start a server on a socket. If the socket is free, Koi starts, otherwise it refuses.

This commit replaces the server mechanic with a 8byte shared memory block. Only 1 single Koi instance can have exclusive access to it. If the Koi instance cannot get exclusive access to the shared memory block, it will not start. If Koi quits normally or crashes, the OS will free up the shared memory access.

Since we can remove the whole QNetwork dependency + do not have to listen to a socket, the memory usage is further reduces by ~2MB (mostly due to not loading libQt6Network.so).

It is also unlikely, that QNetwork will be required by Koi in the near future.